criacao da DSL, alteracao da anotacao que identifica as linhas, voltei a lancar FFPojoException para manter a retrocompatibilidade, mas alterei para que ele herde de RuntimeException e deixe de ser excecao checked
