+++ ffpojo/src/main/java/com/github/ffpojo/FFPojoHelper.java
@@ -2,7 +2,7 @@ package com.github.ffpojo;
 
 import com.github.ffpojo.container.HybridMetadataContainer;
 import com.github.ffpojo.container.MetadataContainer;
-import com.github.ffpojo.exception.FFPojoRuntimeException;
+import com.github.ffpojo.exception.FFPojoException;
 import com.github.ffpojo.exception.MetadataContainerException;
 import com.github.ffpojo.metadata.RecordDescriptor;
 import com.github.ffpojo.parser.RecordParser;
@@ -15,48 +15,48 @@ public class FFPojoHelper {
 	
 	private MetadataContainer metadataContainer;
 	
-	private FFPojoHelper() throws FFPojoRuntimeException {
+	private FFPojoHelper() throws FFPojoException {
 		try {
 			this.metadataContainer = new HybridMetadataContainer();
 		} catch (MetadataContainerException e) {
-			throw new FFPojoRuntimeException(e);
+			throw new FFPojoException(e);
 		}
 	}
 	
-	public static FFPojoHelper getInstance() throws FFPojoRuntimeException {
+	public static FFPojoHelper getInstance() throws FFPojoException {
 		if (singletonInstance == null) {
 			singletonInstance = new FFPojoHelper();
 		}
 		return singletonInstance;
 	}
 	
-	public <T> T createFromText(Class<T> recordClazz, String text) throws FFPojoRuntimeException {
+	public <T> T createFromText(Class<T> recordClazz, String text) throws FFPojoException {
 		try{
 			RecordDescriptor recordDescriptor = metadataContainer.getRecordDescriptor(recordClazz);
 			RecordParser recordParser = RecordParserFactory.createRecordParser(recordDescriptor);
 			return recordParser.parseFromText(recordClazz, text);
 		}catch(Exception e){
-			throw new FFPojoRuntimeException(e);
+			throw new FFPojoException(e);
 		}
 		
 	}
 	
-	public <T> String parseToText(T record) throws FFPojoRuntimeException {
+	public <T> String parseToText(T record) throws FFPojoException {
 		try{			
 			RecordDescriptor recordDescriptor = metadataContainer.getRecordDescriptor(record.getClass());
 			RecordParser recordParser = RecordParserFactory.createRecordParser(recordDescriptor);
 			return recordParser.parseToText(record);
 		}catch(Exception e){
-			throw new FFPojoRuntimeException(e);
+			throw new FFPojoException(e);
 		}
 	}
 	
-	public RecordParser getRecordParser(Class<?> recordClazz) throws FFPojoRuntimeException {
+	public RecordParser getRecordParser(Class<?> recordClazz) throws FFPojoException {
 		try{
 			RecordDescriptor recordDescriptor = metadataContainer.getRecordDescriptor(recordClazz);
 			return RecordParserFactory.createRecordParser(recordDescriptor);			
 		}catch(Exception e){
-			throw new FFPojoRuntimeException(e);
+			throw new FFPojoException(e);
 		}
 	}
