+++ ffpojo/src/main/java/com/github/ffpojo/dsl/FFPojoRandomReaderBuilder.java
@@ -1,0 +1,11 @@
+package com.github.ffpojo.dsl;
+
+import java.util.List;
+
+public interface FFPojoRandomReaderBuilder<T> {
+
+	public List<T> randomRead ();
+	
+	public void randomRead(ReadProcessor processor);
+	
+}
