+++ ffpojo/src/main/java/com/github/ffpojo/dsl/ReadProcessor.java
@@ -1,0 +1,7 @@
+package com.github.ffpojo.dsl;
+
+public interface ReadProcessor {
+	
+	void process(Object item);
+
+}
