+++ ffpojo/src/main/java/com/github/ffpojo/metadata/positional/annotation/PositionalRecord.java
@@ -11,5 +11,5 @@ public @interface PositionalRecord {
 	public static final PositionalRecordLineIdentifier[] defaultIdentifier = new PositionalRecordLineIdentifier[]{};
 	boolean ignorePositionNotFound() default false;
 	boolean autoFillRemainPosition() default false;
-	PositionalRecordLineIdentifier recordLineIdentifier() default @PositionalRecordLineIdentifier(startPosition = 0, textIdentifier="");
+	String textLineIdentifier() default "";
 }
