+++ ffpojo/src/main/java/com/github/ffpojo/metadata/positional/annotation/PositionalRecordLineIdentifier.java
@@ -6,8 +6,8 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
+@Target({ElementType.ANNOTATION_TYPE})
 public @interface PositionalRecordLineIdentifier {
-	int startPosition();
+//	int startPosition();
 	String textIdentifier();
 }
